payment-backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.12,<3.0
python-dotenv>=1.0,<2.0
